Clarify exports and imports
When exporting or importing you must take into account the context, since it is not the same to export or import in javascript for node.js than for native javascript. I think the user must understand when an export or a module.exports is used. On the other hand, including the call from the html will help a lot to test the example. I hope to be of help.